requirements.txt
asgiref==3.4.1
Django==3.2.7
gunicorn==20.1.0
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
pkg-resources==0.0.0
PyJWT==1.7.1